+++ src/main/java/com/cedarsoftware/util/io/factory/LocalDateFactory.java
@@ -4,6 +4,8 @@ import com.cedarsoftware.util.io.JsonObject;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.util.Map;
+import java.util.Optional;
 
 
 public class LocalDateFactory extends AbstractTemporalFactory<LocalDate> {