+++ src/main/java/com/cedarsoftware/util/io/factory/LocalTimeFactory.java
@@ -6,6 +6,7 @@ import com.cedarsoftware.util.io.MetaUtils;
 
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import java.util.Map;
 
 
 public class LocalTimeFactory extends AbstractTemporalFactory<LocalTime> {