+++ src/test/java/com/cedarsoftware/util/io/factory/LocalTimeFactoryTests.java
@@ -28,7 +28,7 @@ class LocalTimeFactoryTests {
         var factory = new LocalTimeFactory();
         var jsonObject = buildJsonObject(hour, minute, second, nano);
 
-        LocalTime time = factory.newInstance(LocalTime.class, jsonObject, new HashMap());
+        LocalTime time = (LocalTime) factory.newInstance(LocalTime.class, jsonObject);
 
         assertThat(time).hasHour(hour)
                 .hasMinute(minute)
@@ -56,7 +56,7 @@ class LocalTimeFactoryTests {
         var jsonObject = new JsonObject();
         jsonObject.put("value", "09:27:39+01:00");
 
-        LocalTime time = factory.newInstance(LocalTime.class, jsonObject, new HashMap());
+        LocalTime time = factory.newInstance(LocalTime.class, jsonObject);
 
         assertThat(time).hasHour(9)
                 .hasMinute(27)